+++ src/alice/models/review_card.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 from enum import StrEnum
 
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import Base, TimestampMixin
@@ -24,7 +24,7 @@ class ReviewCard(Base, TimestampMixin):
     __tablename__ = "review_cards"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
     concept_id: Mapped[str] = mapped_column(String(255), nullable=False)
     review_prompt: Mapped[str] = mapped_column(Text, nullable=False)
     # FSRS fields
